requirements.txt
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
xgboost>=2.0.0
lightgbm>=4.0.0
shap>=0.44.0
joblib>=1.3.0
# Database dependencies
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
# Production configuration
python-dotenv>=1.0.0
